requirements.txt
Pillow>=9.0
wxPython>=4.2.0
